pesan 7

+++ database/migrations/2025_12_27_020330_add_payment_status_to_orders_table.php
@@ -6,23 +6,24 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-           $table->string('payment_status')->default('unpaid')->after('status');
+            if (!Schema::hasColumn('orders', 'payment_status')) {
+                $table->string('payment_status')
+                      ->default('unpaid')
+                      ->after('status');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-           $table->dropColumn('payment_status');
+            if (Schema::hasColumn('orders', 'payment_status')) {
+                $table->dropColumn('payment_status');
+            }
         });
     }
 };
+
